+++ RELEASE_NOTES.md
@@ -10,7 +10,8 @@ Until then, deprecated features may be removed in the next minor version.
 
 ### ✨ Features
 
-* `DocumentGroupPresenter` is a new protocol to present and dismiss screens from a `DocumentGroup`.
+* `DocumentGroupInspector` is a new protocol that can be implemented by any type that should be able to inspect the current `DocumentGroup`.
+* `DocumentGroupPresenter` is a new protocol that can be implemented by any type that should be able to present and dismiss views from the current `DocumentGroup`.
 * `DocumentGroupModal` is a `View` protocol that can be implemented by any view that can then be opened as a modal from a `DocumentGroup`. 
 * `DocumentGroup` has new onboarding extensions.
 * `UserDefaults` has new onboarding extensions.